+++ tests/utils/ymlparsers_extra_test.py
@@ -19,7 +19,7 @@ logger = logging.getLogger(__name__)
 
 
 @pytest.fixture
-def ymlparsersExtraFixture(request, mocker):
+def ymlparsersExtraFixture(request, mocker, ymlparsersSetup, ymlparsersCleanup):
     p_isHiYaPyCoAvailable = ymlparsers_extra._isHiYaPyCoAvailable
     p_isJinja2DefaultAvailable = ymlparsers_extra._isJinja2DefaultAvailable
 
@@ -33,7 +33,7 @@ def ymlparsersExtraFixture(request, mocker):
     ymlparsers_extra._isJinja2DefaultAvailable = isJinja2DefaultAvailable
     ymlparsers_extra._init_globals()
 
-    yield None
+    yield request_param
     ymlparsers_extra._isHiYaPyCoAvailable = p_isHiYaPyCoAvailable
     ymlparsers_extra._isJinja2DefaultAvailable = p_isJinja2DefaultAvailable
     ymlparsers_extra._init_globals()
@@ -83,6 +83,14 @@ def test_uconvert_template_to_string_format_undocumented2(request, mocker):
     pytest.assume(exp_value == value)
 
 
+def test_uconvert_is_used_in_convert_template_to_string_format(request, mocker):
+    logger.info(f'{request._pyfuncitem.name}()')
+
+    uconvert_mock=mocker.patch('alexber.utils.ymlparsers_extra._convert_template_to_string_format', autospec=True, spec_set=True)
+    convert_template_to_string_format('{{name}}')
+    pytest.assume(uconvert_mock.call_count > 0)
+
+
 #isHiYaPyCoAvailable,isJinja2DefaultAvailable
 @pytest.mark.parametrize('ymlparsersExtraFixture', [(False, False), #minimal, no 3-rd party dependencies
                                                     (True, True),   #maximal 3-rd party dependencies available, \
@@ -91,7 +99,7 @@ def test_uconvert_template_to_string_format_undocumented2(request, mocker):
                                                     (True, False),  #can't really happen
                                                     ], indirect=True)
 def test_convert_template_to_string_format_minimal(request, mocker, ymlparsersExtraFixture):
-    logger.info(f'{request._pyfuncitem.name}()')
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
     exp_value = '{name}'
     template = '{{name}}'
@@ -104,7 +112,7 @@ def test_convert_template_to_string_format_minimal(request, mocker, ymlparsersEx
                                                     ], indirect = True)
 @pytest.mark.yaml
 def test_convert_template_to_string_jinja2DefaultChanged(request, mocker, ymlparsersExtraFixture):
-    logger.info(f'{request._pyfuncitem.name}()')
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
 
     mocker.patch('jinja2.defaults.VARIABLE_START_STRING', new='1_', spec_set=True)
@@ -120,8 +128,8 @@ def test_convert_template_to_string_jinja2DefaultChanged(request, mocker, ymlpar
 @pytest.mark.parametrize('ymlparsersExtraFixture', [(True, False)
                                                     ], indirect = True)
 @pytest.mark.yaml
-def test_convert_template_to_string_HiYaPyCoDefault(request, mocker, ymlparsersExtraFixture, ymlparsersSetup, ymlparsersCleanup):
-    logger.info(f'{request._pyfuncitem.name}()')
+def test_convert_template_to_string_HiYaPyCoDefault(request, mocker, ymlparsersExtraFixture):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
     ymlparsers.initConfig()
 
@@ -150,8 +158,8 @@ def test_convert_template_to_string_HiYaPyCoDefault(request, mocker, ymlparsersE
 @pytest.mark.parametrize('ymlparsersExtraFixture', [(True, False)
                                                     ], indirect = True)
 @pytest.mark.yaml
-def test_convert_template_to_string_HiYaPyCoDefaultChanged(request, mocker, ymlparsersExtraFixture, ymlparsersSetup, ymlparsersCleanup):
-    logger.info(f'{request._pyfuncitem.name}()')
+def test_convert_template_to_string_HiYaPyCoDefaultChanged(request, mocker, ymlparsersExtraFixture):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
 
     ymlparsers.initConfig(jinja2ctx={'variable_start_string':'2_', 'variable_end_string':'_2' })
@@ -177,11 +185,10 @@ def test_convert_template_to_string_HiYaPyCoDefaultChanged(request, mocker, ymlp
     pytest.assume(mock_variable_end_string.call_count > 0)
 
 @pytest.mark.yaml
-def test_convert_template_to_string_explicit_param1(request, mocker, ymlparsersExtraFixture):
-    logger.info(f'{request._pyfuncitem.name}()')
+def test_convert_template_to_string_explicit_param1_jinja2ctx(request, mocker, ymlparsersExtraFixture):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
     mock_lock = create_mock_lock(mocker)
-    mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)
 
     exp_value = '{name}'
     template = '3_name_3'
@@ -189,31 +196,99 @@ def test_convert_template_to_string_explicit_param1(request, mocker, ymlparsersE
     jinja2ctx = _Environment(undefined=_DebugUndefined, variable_start_string='3_', variable_end_string='_3')
     value = convert_template_to_string_format(template, jinja2ctx=jinja2ctx)
     pytest.assume(exp_value == value)
+
+@pytest.mark.yaml
+def test_convert_template_to_string_explicit_param1_jinja2Lock(request, mocker, ymlparsersExtraFixture):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
+
+    mock_lock = create_mock_lock(mocker)
+    mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)
+
+    #exp_value = '{name}'
+    template = '{{name}}'   #HiYaPyCo.jinja2ctx is default one
+
+    jinja2Lock_param_lock = create_mock_lock(mocker)
+    with pytest.raises(ValueError):
+        convert_template_to_string_format(template, jinja2Lock=jinja2Lock_param_lock)
+    #pytest.assume(exp_value == value)
     pytest.assume(mock_lock.acquire.call_count == 0)
+    #pytest.assume(jinja2Lock_param_lock.acquire.call_count ==0)
+
+
+
+#isHiYaPyCoAvailable,isJinja2DefaultAvailable
+@pytest.mark.parametrize('ymlparsersExtraFixture', [(False, True)   #only Jinja2 avaialble \
+                                                    ], indirect = True)
+@pytest.mark.yaml
+def test_convert_template_to_string_explicit_param1a_jinja2Lock(request, mocker, ymlparsersExtraFixture,
+                                                                ):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
+
+    mock_lock = create_mock_lock(mocker)
+    mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)
+    jinja2ctx_mock = mocker.patch.object(HiYaPyCo, 'jinja2ctx', spec_set=True)
+
+    exp_value = '{name}'
+    template = '{{name}}'   #isHiYaPyCoAvailable is not available, global default are used and external lock
+                            # (will be ignored)
+
+    jinja2Lock_param_lock = create_mock_lock(mocker)
+    value = convert_template_to_string_format(template, jinja2Lock=jinja2Lock_param_lock)
+    pytest.assume(exp_value == value)
+    pytest.assume(mock_lock.acquire.call_count == 0)
+    pytest.assume(jinja2ctx_mock.call_count == 0)
+
+@pytest.mark.yaml
+def test_convert_template_to_string_explicit_param1b_jinja2Lock(request, mocker, ymlparsersExtraFixture,
+                                                               ):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
+
+    ymlparsers.initConfig(jinja2ctx={'variable_start_string': '22_', 'variable_end_string': '_22'})
+
+    init_jinja2ctx = HiYaPyCo.jinja2ctx
+    mock_lock = create_mock_lock(mocker)
+    mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)
+    jinja2ctx_mock = mocker.patch.object(HiYaPyCo, 'jinja2ctx', spec_set=True)
+    mock_variable_start_string = mocker.PropertyMock(return_value=init_jinja2ctx.variable_start_string)
+    type(jinja2ctx_mock).variable_start_string = mock_variable_start_string
+    mock_variable_end_string = mocker.PropertyMock(return_value=init_jinja2ctx.variable_end_string)
+    type(jinja2ctx_mock).variable_end_string = mock_variable_end_string
+
+    #exp_value = '{name}'
+    template = '22_name_22'   #HiYaPyCo.jinja2ctx is non default one and external lock
+
+    jinja2Lock_param_lock = create_mock_lock(mocker)
+    with pytest.raises(ValueError):
+        convert_template_to_string_format(template, jinja2Lock=jinja2Lock_param_lock)
+    #pytest.assume(exp_value == value)
+    pytest.assume(mock_lock.acquire.call_count == 0)
+    #pytest.assume(jinja2Lock_param_lock.acquire.call_count ==0)
+
 
 #isHiYaPyCoAvailable,isJinja2DefaultAvailable
 @pytest.mark.parametrize('ymlparsersExtraFixture', [(False, True)
                                                     ], indirect = True)
 @pytest.mark.yaml
-def test_convert_template_to_string_explicit_param1a(request, mocker, ymlparsersExtraFixture):
-    logger.info(f'{request._pyfuncitem.name}()')
-    from jinja2 import Environment as _Environment, \
-        DebugUndefined as _DebugUndefined
+def test_convert_template_to_string_explicit_param1a_jinja2ctx(request, mocker, ymlparsersExtraFixture,
+                                                               ):
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
     mock_lock = create_mock_lock(mocker)
     mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)
 
     exp_value = '{name}'
-    template = '3_name_3'
+    template = '33_name_33'
 
-    jinja2ctx = _Environment(undefined=_DebugUndefined, variable_start_string='3_', variable_end_string='_3')
+    jinja2ctx = _Environment(undefined=_DebugUndefined, variable_start_string='33_', variable_end_string='_33')
     value = convert_template_to_string_format(template, jinja2ctx=jinja2ctx)
+
     pytest.assume(exp_value == value)
     pytest.assume(mock_lock.acquire.call_count == 0)
 
+
 @pytest.mark.yaml
 def test_convert_template_to_string_explicit_param2(request, mocker, ymlparsersExtraFixture):
-    logger.info(f'{request._pyfuncitem.name}()')
+    logger.info(f'{request._pyfuncitem.name}{ymlparsersExtraFixture}')
 
     mock_lock = create_mock_lock(mocker)
     mocker.patch.object(HiYaPyCo, 'jinja2Lock', new=mock_lock, spec_set=True)